requirements.txt
Flask
yfinance
line-bot-sdk
